+++ packages/account-core/src/index.ts
@@ -1,6 +1,18 @@
 export * from "./config";
 export * from "./wallet";
+export {
+  connect,
+  disconnect,
+  onEvent,
+  switchNetwork,
+  getWalletNetwork,
+  getSupportedEthereumNetworks,
+  getEthereumWallets,
+} from "@dynamic-labs/global-wallet-client/features";
+export { createEIP1193Provider } from "@dynamic-labs/global-wallet-client/ethereum";
 export * from "./sessionHelper";
 export * from "./abis";
+export * from "./snsHelper";
 export { isEthereumWallet } from "@dynamic-labs/ethereum";
 export { isZKsyncConnector } from "@dynamic-labs/ethereum-aa-zksync";
+export { SiwsMessage } from "./siws";