+++ packages/account-core/tsup.config.ts
@@ -3,13 +3,8 @@ import path from "path";
 
 export default defineConfig((options) => {
   return {
-    entry: [
-      "src/index.ts",
-      "src/wallet.ts",
-      "src/config.ts",
-      "src/sessionHelper.ts",
-      "src/abis/*",
-    ],
+    entry: ["src/index.ts", "src/wallet.ts", "src/config.ts", "src/siws.ts", "src/sessionHelper.ts",
+      "src/abis/*"],
     splitting: true,
     sourcemap: !!options.watch,
     clean: true,