+++ packages/account-eip6963/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sophon-labs/account-eip6963
 
+## 1.1.13
+
+### Patch Changes
+
+- Exposing more methods
+- Updated dependencies
+  - @sophon-labs/account-core@1.1.13
+
+## 1.1.12
+
+### Patch Changes
+
+- Updating dependencies and exposing global wallet native api
+- Updated dependencies
+  - @sophon-labs/account-core@1.1.12
+
 ## 1.1.11
 
 ### Patch Changes